+++ src/phoenix/experiments/evaluators/__init__.py
@@ -1,18 +1,18 @@
-from phoenix.datasets.evaluators.code_evaluators import (
+from phoenix.experiments.evaluators.code_evaluators import (
     ContainsAllKeywords,
     ContainsAnyKeyword,
     ContainsKeyword,
     JSONParsable,
     MatchesRegex,
 )
-from phoenix.datasets.evaluators.llm_evaluators import (
+from phoenix.experiments.evaluators.llm_evaluators import (
     CoherenceEvaluator,
     ConcisenessEvaluator,
     HelpfulnessEvaluator,
     LLMCriteriaEvaluator,
     RelevanceEvaluator,
 )
-from phoenix.datasets.evaluators.utils import create_evaluator
+from phoenix.experiments.evaluators.utils import create_evaluator
 
 __all__ = [
     "create_evaluator",
